+++ CHANGELOG.md
@@ -5,7 +5,9 @@
   Configured with python.minVirtualenvVersion. Important because v20 changes pip installation
     strategy in environment to the same version as bundled with virtualenv instead of always
     downloading new pip (extremely important for python 2 support because pip 21 does not support it anymore)
-* Fix up-to-date check in pipInstall task for VCS modules (the latest pip changed freeze output syntax)  
+* Fix up-to-date check in pipInstall task for VCS modules (the latest pip changed freeze output syntax)
+* Hide passwords in logs for pip extra index repository urls with credentials (extraIndexUrls config) (#12)
+* Add general logged command cleaning mechanism to Python object for plugin authors (Pip object uses it for passwords hiding)
 
 ### 2.2.0 (2020-04-06)
 * Add support for `pip` flags: [--trusted-host](https://pip.pypa.io/en/stable/reference/pip/#trusted-host)